% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcrpheno_clones}
\alias{tcrpheno_clones}
\title{A list of 8 single-cell T cell receptor sequences runs.}
\description{
A data frame of 8 \code{filtered_contig_annotations.csv} files
outputted from 10X Cell Ranger. The aligned data was then processed by scRepertoire
and exported using the exportClones() function. More information on the
data can be found in the following \href{https://pubmed.ncbi.nlm.nih.gov/39024572/}{manuscript}.
}
\concept{Data}
