% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_tcrs.R
\name{score_tcrs}
\alias{score_tcrs}
\title{score TCR sequences}
\usage{
score_tcrs(data, chain)
}
\arguments{
\item{data}{TCR sequence data}

\item{chain}{whether the TCR data is paired a-b ("ab"), alpha only ("a") or beta only ("b")}
}
\description{
score TCR sequences
}
